requirements.txt
attrdict==2.0.0
chardet==3.0.4
click==6.7
cytoolz==0.9.0.1
eth-abi==1.1.1
eth-account==0.2.2
eth-hash==0.1.3
eth-keyfile==0.5.1
eth-keys==0.2.0b3
eth-rlp==0.1.2
eth-utils==1.0.3
Flask==1.0.2
hexbytes==0.1.0
idna==2.6
itsdangerous==0.24
Jinja2==2.10
lru-dict==1.1.6
MarkupSafe==1.0
parsimonious==0.8.0
py-solc==3.0.0
pycryptodome==3.6.1
requests==2.18.4
rlp==1.0.1
semantic-version==2.6.0
six==1.11.0
toolz==0.9.0
urllib3==1.22
web3==4.3.0
websockets==4.0.1
Werkzeug==0.14.1
